+++ gapic/schema/api.py
@@ -378,6 +378,7 @@ class API:
     all_protos: Mapping[str, Proto]
     service_yaml_config: service_pb2.Service
     subpackage_view: Tuple[str, ...] = dataclasses.field(default_factory=tuple)
+    gapic_version: str = "0.0.0"
 
     @classmethod
     def build(
@@ -493,12 +494,16 @@ class API:
         ParseDict(
             opts.service_yaml_config, service_yaml_config, ignore_unknown_fields=True
         )
+        gapic_version = opts.gapic_version
 
         # Third pass for various selective GAPIC settings; these require
         # settings in the service.yaml and so we build the API object
         # before doing another pass.
         api = cls(
-            naming=naming, all_protos=protos, service_yaml_config=service_yaml_config
+            naming=naming,
+            all_protos=protos,
+            service_yaml_config=service_yaml_config,
+            gapic_version=gapic_version,
         )
 
         if package in api.all_library_settings:
@@ -554,6 +559,7 @@ class API:
                     naming=naming,
                     all_protos=new_all_protos,
                     service_yaml_config=service_yaml_config,
+                    gapic_version=gapic_version,
                 )
 
         return api
@@ -930,11 +936,13 @@ class API:
 
             # Check to see if selective gapic generation methods are valid.
             selective_gapic_errors = {}
+            # TODO(https://github.com/googleapis/gapic-generator-python/issues/2446):
+            # Workaround issue in Python 3.14 related to code coverage by adding `# pragma: no branch`
             for (
                 method_name
             ) in (
                 library_settings.python_settings.common.selective_gapic_generation.methods
-            ):
+            ):  # pragma: no branch
                 if method_name not in self.all_methods:
                     selective_gapic_errors[method_name] = "Method does not exist."
                 elif not method_name.startswith(library_settings.version):