+++ gapic/utils/options.py
@@ -50,6 +50,7 @@ class Options:
     service_yaml_config: Dict[str, Any] = dataclasses.field(default_factory=dict)
     rest_numeric_enums: bool = False
     proto_plus_deps: Tuple[str, ...] = dataclasses.field(default=("",))
+    gapic_version: str = "0.0.0"
 
     # Class constants
     PYTHON_GAPIC_PREFIX: str = "python-gapic-"
@@ -71,6 +72,7 @@ class Options:
             # proto plus dependencies delineated by '+'
             # For example, 'google.cloud.api.v1+google.cloud.anotherapi.v2'
             "proto-plus-deps",
+            "gapic-version",  # A version string following https://peps.python.org/pep-0440
         )
     )
 
@@ -165,6 +167,16 @@ class Options:
             "TRUE",
         )
 
+        # `rest-numeric-enums` is False by default. Make sure users can also disable
+        # it by passing `rest-numeric-enums=False`.
+        rest_numeric_enums = opts.pop("rest-numeric-enums", ["False"])[0] in (
+            "True",
+            "true",
+            "T",
+            "t",
+            "TRUE",
+        )
+
         # NOTE: Snippets are not currently correct for the alternative (Ads) templates
         # so always disable snippetgen in that case
         # https://github.com/googleapis/gapic-generator-python/issues/1052
@@ -195,8 +207,9 @@ class Options:
             # transport should include desired transports delimited by '+', e.g. transport='grpc+rest'
             transport=opts.pop("transport", ["grpc"])[0].split("+"),
             service_yaml_config=service_yaml_config,
-            rest_numeric_enums=bool(opts.pop("rest-numeric-enums", False)),
+            rest_numeric_enums=rest_numeric_enums,
             proto_plus_deps=proto_plus_deps,
+            gapic_version=opts.pop("gapic-version", ["0.0.0"]).pop(),
         )
 
         # Note: if we ever need to recursively check directories for sample